requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
